added .aliases

+++ bootstrap.sh
@@ -8,6 +8,9 @@ force=0
 # built in variables
 files2sync="
   .bashrc 
+  .bash_profile
+  .aliases
+  .profile
   .inputrc 
   .tmux.conf 
   .tmuxStatus.sh
@@ -76,10 +79,13 @@ function report_existing(){
 function create_symlinks(){
   for f in "$@"; do
     TARGET=$HOME/$f
-    if [ -e $TARGET ]; then
-      rm $TARGET
+    SOURCE=$PWD/$f
+    if [ -e $SOURCE ]; then
+      if [ -e $TARGET ]; then
+        rm $TARGET
+      fi
+      ln -s -t $HOME $SOURCE
     fi
-    ln -s -t $HOME $PWD/$f
   done
 }
 
@@ -88,7 +94,6 @@ function create_symlinks(){
 parse_args "$@" 
 any_exists $files2sync # set variable $exists
 if [ "$exists" == "0" -o "$force" == "1" ]; then
-  echo "going to make links"
   create_symlinks $files2sync
 else
   report_existing $files2sync
